refactor: remove residual modal/api coupling from the local app
The runtime was already backend-neutral; this clears the leftover
artifacts that still named or branched on the backend:

- Drop the modal/api runner badge and runnerFromId() prefix matching
  (official-plugins.server.ts, plugins-dialog.tsx); the plugin manager
  no longer classifies plugins by backend.
- Remove the stale combine-text texts->text adapter from both the TS
  runner (runners/generic.ts) and the SDK engine (engine/invoker.py).
  The gemini plugin already reads input.texts per the ABI; the adapter
  conflicted with it.
- Delete dead ModalPluginPayload type (use-plugins-registry.ts).
- Rename convert-modal-output-fileref.ts -> convert-output-fileref.ts
  and update all references (TS import, CLAUDE.md, docs, SDK comments).
- Neutralize tongflow-modal-* examples in comments/placeholders.

Housekeeping bundled in: fix CLAUDE.md (src/lib subdirs, workflowId?),
gitignore Python tool caches, drop a duplicate PluginRunner type.
